requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
pymongo
redis
openai
google-cloud-firestore # Assuming Firebase interaction might use Firestore
requests
python-multipart # For file uploads

# Document processing libraries
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
beautifulsoup4==4.12.2
markdown==3.4.3
pyyaml==6.0.1

# External services
sendgrid==6.10.0  # For SendGrid email integration
boto3==1.28.38    # For AWS S3 storage integration

# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-asyncio==0.21.1  # For testing async code
pytest-httpx==0.23.0    # For testing HTTP clients
aioresponses==0.7.4     # For mocking async HTTP requests
